+++ src/CHANGELOG.md
@@ -1,6 +1,6 @@
 [0.1.0]
 -------
-- [ ] Add card deletion
 - [ ] Implement actual PWA behavior
+- [ ] Add card management modal
 - [ ] Implement game session
 - [x] Make images part of SVGs